Ensure redirect and query params input are merged
Fixes an issue where, if the `redirect.to` value contained a query string, and the incoming `input.params` were present, two query string values were being appended.

For example, if the `to` value was `/foo?x=y` and the incoming input params were `{ a: 'b' }`, the final `to` value became `/foo?x=y?a=b`. This subsequently caused bad requests when the `to` value was accessed.

The above example will now return as `/foo?x=y&a=b`.

If there is a conflicting query param key, the value of the redirect wins. For example:
With a `to` value of `/foo?x=y` and an input of `{ x: 'z' }`, the final result would be `/foo?x=y` _not `/foo?x=z`.