GDPR/Privacy Compliance added to inquiry form
EBM sites need a privacy compliance checkbox added to the inquiry forms, so if the user selects one of the affected countries, a checkbox will appear with a custom label set in site.js (`gdprMessage`).  By default, the message will be null and the checkbox (`gdprOptIn`) will be set to false.

Additionally, a second message is being added at the bottom of the form, regardless of country, which can also be set in site.js (`privacyMessage`).